+++ packages/plugin-audio/web/src/editor.ts
@@ -1,5 +1,4 @@
 import { createAudioData } from './createAudioData';
-
 import { createAudioPlugin } from './createAudioPlugin';
 import type { AudioPluginEditorConfig } from './types';
 import { AUDIO_TYPE } from './types';
@@ -8,6 +7,7 @@ import type { EditorPluginCreator } from 'wix-rich-content-common';
 import { tiptapExtensions } from './tiptap/tiptap';
 import type { TiptapEditorPlugin } from 'ricos-tiptap-types';
 import { getToolbarButtons } from './getToolbarButtons';
+import { getAddButtons } from './getAddButtons';
 
 export const pluginAudio: EditorPluginCreator<AudioPluginEditorConfig> = config => {
   const pluginConfig = { ...DEFAULTS.config, ...config };
@@ -18,5 +18,6 @@ export const pluginAudio: EditorPluginCreator<AudioPluginEditorConfig> = config 
     createPluginData: createAudioData,
     tiptapExtensions,
     toolbarButtons: getToolbarButtons(config),
+    addButtons: getAddButtons(config),
   } as TiptapEditorPlugin;
 };